requirements.txt
PyYAML
requests
psycopg2-binary
arelle-release
